+++ config/effective_orders.rb
@@ -238,9 +238,17 @@ EffectiveOrders.setup do |config|
 
   # Stripe configuration
   config.stripe_enabled = false
-  config.stripe_subscriptions_enabled = false # https://stripe.com/docs/subscriptions
+
   config.stripe_connect_enabled = false # https://stripe.com/docs/connect
   config.stripe_connect_application_fee_method = Proc.new { |order_item| order_item.total * 0.10 } # 10 percent
+
+  config.subscriptions_enabled = false # https://stripe.com/docs/subscriptions
+
+  config.subscription = {
+    trial_name: 'Free Trial',
+    trial_description: '45-Day Free Trial',
+    trial_period: 45.days
+  }
 
   if Rails.env.production?
     config.stripe = {
